\name{FemPYSweden}
\alias{FemPYSweden}
\docType{data}
\title{Population estimate for females in Sweden}
\description{
Population estimates for females in Sweden aged 25--84 years for single calendar years from 1951--2010.
}
\usage{data(FemPYSweden)}
\format{
  A data frame with 720 observations provided in 60 rows and 12 columns. Each row provides
  data for one period (year) and all age groups. Each row contains data for one age group 
  and all periods.
}
%%\details{}
\source{WHO Mortality database, http://www.who.int/whosis/mort/download/en/index.html}
\examples{
data(FemPYSweden)
}
\keyword{datasets}
