\name{plotBAPC}
\alias{plotBAPC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
    Function to plot age-specific and/or agestandardized rates and projections of a Bayesian age-period-cohort model
}
\description{
Function to plot age-specific and/or agestandardized projections of a Bayesian age-period-cohort model
generated using \code{BAPC}.
}
\usage{
plotBAPC(APCList, scale = 10^3, type="ageSpecProj", 
    showdata = FALSE, mfrow = NULL, coladd = NULL, ...)
}
\arguments{
  \item{APCList}{
     \code{APCList} object generated by the \code{BAPC} function.
}
  \item{scale}{
    Scaling factor for the incidence/mortality rates.
}
  \item{type}{
   String indicating what should be plotted: 
   \code{ageSpecRate} --- projected age-specific rates,
   \code{ageSpecProj} --- projected age-specific counts, 
   \code{ageSpecBoth} --- projected age-specific rates and counts in one plot, 
   \code{ageStdRate} --- projected age-standardized rates,
   \code{ageStdProj} --- projected age-standardized counts or
   \code{ageStdProj} --- projected age-standardized rates and counts in one plot.
}
  \item{showdata}{
    Boolean indicating whether observed data points should be added to the plot
}
  \item{mfrow}{
    A vector of the form 'c(nr, nc)'. Subsequent figures will be drawn in 
    an 'nr'-by-'nc' array on the device by _columns_ ('mfcol'), 
    or _rows_ ('mfrow'), respectively. 
}
  \item{coladd}{
    color used for intervals of projected counts when \code{type=="ageSpecBoth"} or \code{type=="ageStdBoth"}.
}
  \item{\dots}{
    Further arguments passed to \code{par()}
}
}

\value{
    One or two plotting devices illustrating age-specific and/or age-standardized 
    projections/fitted values.
}

\author{
    Andrea Riebler
}

\examples{
\dontrun{
if(requireNamespace("INLA", quietly = TRUE)) {
    require(INLA)
    data(FemLCSweden)
    data(FemPYSweden)
    data(whostandard)

    lc_sweden <- APCList(FemLCSweden, FemPYSweden, gf=5)

    lc_sweden <- BAPC(lc_sweden, predict=list(npredict=10, retro=TRUE), 
        verbose=FALSE, stdweight=whostandard[6:17,2])

    plotBAPC(lc_sweden, scale=10^5, type="ageSpecProj", showdata=TRUE)

    ## Assume the data for 2000 to 2010 are not observed 
    ## (projection results are the same as above)
    FemLCSweden[51:60,] <- NA

    lc_sweden <- APCList(FemLCSweden, FemPYSweden, gf=5)

    lc_sweden <- BAPC(lc_sweden, predict=list(npredict=10, retro=FALSE), 
        verbose=FALSE, stdweight=whostandard[6:17,2])

    plotBAPC(lc_sweden, scale=10^5, type="ageSpecProj", showdata=TRUE)
}
}
}

