\name{qapc}
\alias{qapc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Function to compute quantiles for age-specific (and age-standardized) rates or projections.
}
\description{
Function to compute quantiles for age-specific (and age-standardized)
rates or projections.
}
\usage{
  qapc(APCList, percentiles=c(0.025, seq(0.05, 0.95, by=0.05), 0.975))
}
\arguments{
\item{APCList}{
     \code{APCList} object generated by the \code{BAPC} function.
}
\item{percentiles}{
   Probabilities related to percentiles or prediction intervals
          which should be computed.
          These values must be between 0 and 1.
}
}

\value{
  An APCList object where the desired quantiles are added to
  \code{agespec.rate(APCList)}, \code{agespec.proj(APCList)} and if
  available  \code{agestd.rate(APCList)}, \code{agestd.proj(APCList)}. 
  Note only quantiles that are not yet available will be added.
}

\author{
    Andrea Riebler
}

\examples{
\dontrun{
if(requireNamespace("INLA", quietly = TRUE)) {
    require(INLA)
    data(FemLCSweden)
    data(FemPYSweden)
    data(whostandard)

    lc_sweden <- APCList(FemLCSweden, FemPYSweden, gf=5)

    result <- BAPC(lc_sweden, predict=list(npredict=10, retro=TRUE),
        verbose=FALSE, stdweight=whostandard[6:17,2])

    result <- qapc(result)
    head(agespec.proj(result)[[1]])
}
}
}

